Fix logic in MemoryGame component
In useEffect in Cardboard.tsx, there was a flag that level was not in the dependency array but was used inside the effect. This led to the choice to move level logic to the parent MemoryGame component, which already managed the level state. This way, level changes would trigger the effect correctly.